Fix CI lint regressions in prompt editor
