+++ ui/src/features/header/dashboard-session-tabs.tsx
@@ -1,50 +1,46 @@
 import {
-  type ChangeEvent,
-  type FormEvent,
   type KeyboardEvent as ReactKeyboardEvent,
   useEffect,
   useId,
   useRef,
 } from "react";
-
 import type { DashboardStreamState } from "../../api/dashboard/types";
-import {
-  type FactorySessionSummary,
-  type FactorySessionTarget,
-  FactorySessionsAPIError,
-} from "../../api/factory-sessions";
+import type { FactorySessionSummary } from "../../api/factory-sessions";
 import { DEFAULT_FACTORY_SESSION_ID } from "../../api/session-routing";
-import { Button, Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, Input } from "../../components/ui";
+import { Button, Dialog } from "../../components/ui";
 import { cn } from "../../lib/cn";
 import { DASHBOARD_BODY_TEXT_CLASS } from "../../components/ui/dashboard-typography";
 import { useDashboardStreamStore } from "../dashboard/state/dashboardStreamStore";
 import { DashboardHeaderActionButton } from "./components/dashboard-header-action-button";
+import { OpenSessionDialog } from "./dashboard-session-tabs-open-dialog";
+import {
+  normalizeFactorySessionsError,
+  sessionCloseLabel,
+  sessionPanelID,
+  sessionStreamStatusLabel,
+  sessionTabID,
+  sessionTabLabel,
+} from "./dashboard-session-tabs-utils";
 import { getHeaderControlsMessages } from "./messages/header-controls";
 import {
   type DashboardSessionTabsState,
   useDashboardSessionTabsState,
 } from "./use-dashboard-session-tabs-state";
-
 const SESSION_TABS_SHELL_CLASS = "grid min-w-0 flex-1 gap-2";
 const SESSION_TABS_ROW_CLASS = "flex min-w-0 items-center gap-1.5";
 const SESSION_TAB_LIST_CLASS = "flex min-w-0 flex-1 items-end gap-1 overflow-x-auto pb-1";
-const SESSION_SECTION_LABEL_CLASS =
-  "text-xs uppercase tracking-[0.18em] text-af-ink/52";
 const SESSION_TAB_ITEM_CLASS =
   "group relative flex min-w-0 shrink-0 items-stretch rounded-t-xl rounded-b-md border border-b-0 transition-colors";
 const SESSION_TAB_BUTTON_CLASS =
-  "min-w-0 flex-1 rounded-l-[inherit] px-3 py-2 text-left focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-af-accent/25";
+  "min-w-0 flex-1 rounded-bl-md rounded-tl-xl px-3 py-2 text-left focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-af-accent/25";
 const SESSION_TAB_CLOSE_BUTTON_CLASS =
-  "rounded-r-[inherit] px-2.5 text-sm font-semibold focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-af-accent/25";
+  "rounded-br-md rounded-tr-xl px-2.5 text-sm font-semibold focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-af-accent/25";
 const SESSION_TAB_ACTIVE_CLASS =
   "z-10 border-af-overlay/14 bg-af-surface/72 text-af-ink shadow-[0_-1px_0_rgba(255,255,255,0.08)_inset]";
 const SESSION_TAB_INACTIVE_CLASS =
   "border-af-overlay/14 bg-af-surface/72 text-af-ink/74 hover:border-af-overlay/18 hover:bg-af-surface/72 hover:text-af-ink";
 const SESSION_DIALOG_ERROR_CLASS =
   "rounded-xl border border-af-danger/32 bg-af-danger/8 px-3 py-2 text-sm text-af-ink";
-const SESSION_TARGET_LIST_CLASS = "grid gap-2 sm:grid-cols-2";
-const SESSION_TARGET_BUTTON_CLASS =
-  "flex min-h-11 flex-col items-start justify-center rounded-xl border border-af-overlay/12 bg-af-overlay/4 px-3 py-2 text-left text-sm text-af-ink/82 transition-colors hover:border-af-accent/30 hover:bg-af-overlay/8 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-af-accent/25";
 
 export function DashboardSessionTabs({
   hideOpenButton = false,
@@ -55,11 +51,31 @@ export function DashboardSessionTabs({
   locale: string;
   state?: DashboardSessionTabsState;
 }) {
+  const fallbackState = useDashboardSessionTabsState();
+  const sessionTabsState = state ?? fallbackState;
+
+  return (
+    <DashboardSessionTabsView
+      hideOpenButton={hideOpenButton}
+      locale={locale}
+      state={sessionTabsState}
+    />
+  );
+}
+
+function DashboardSessionTabsView({
+  hideOpenButton,
+  locale,
+  state,
+}: {
+  hideOpenButton: boolean;
+  locale: string;
+  state: DashboardSessionTabsState;
+}) {
   const messages = getHeaderControlsMessages(locale);
   const streamStatus = useDashboardStreamStore(
     (state) => state.streamState.status,
   );
-  const sessionTabsState = state ?? useDashboardSessionTabsState();
   const {
     activeSession,
     activeSessionID,
@@ -79,7 +95,7 @@ export function DashboardSessionTabs({
     setActiveSessionID,
     setDialogOpen,
     setFolderPath,
-  } = sessionTabsState;
+  } = state;
 
   useEffect(() => {
     if (sessions.length === 0) {
@@ -408,191 +424,6 @@ function SessionTabButton({
   );
 }
 
-function OpenSessionDialog({
-  dialogError,
-  discoveredTargets,
-  folderPath,
-  isPending,
-  messages,
-  onChangeFolderPath,
-  onInspectFolder,
-  onOpenTarget,
-}: {
-  dialogError: FactorySessionsAPIError | null;
-  discoveredTargets: FactorySessionTarget[];
-  folderPath: string;
-  isPending: boolean;
-  messages: ReturnType<typeof getHeaderControlsMessages>;
-  onChangeFolderPath: (value: string) => void;
-  onInspectFolder: (event: FormEvent<HTMLFormElement>) => void;
-  onOpenTarget: (target: FactorySessionTarget) => void;
-}) {
-  const folderFieldID = useId();
-  const folderPickerInputRef = useRef<HTMLInputElement | null>(null);
-
-  async function handleOpenFolderPicker() {
-    const directoryHandle = await pickDirectoryHandle();
-    if (directoryHandle) {
-      const selectedPath = readDirectoryHandlePath(directoryHandle);
-      if (selectedPath) {
-        onChangeFolderPath(selectedPath);
-        return;
-      }
-    }
-
-    folderPickerInputRef.current?.click();
-  }
-
-  function handleSelectFolder(event: ChangeEvent<HTMLInputElement>) {
-    const nextFolderPath = extractSelectedFolderPath(event.target.files);
-    if (nextFolderPath) {
-      onChangeFolderPath(nextFolderPath);
-    }
-    event.target.value = "";
-  }
-
-  return (
-    <DialogContent closeDisabled={isPending}>
-      <DialogHeader>
-        <DialogTitle>{messages.openSessionDialogTitle}</DialogTitle>
-        <DialogDescription>
-          {messages.openSessionDialogDescription}
-        </DialogDescription>
-      </DialogHeader>
-      <form className="grid gap-4" onSubmit={onInspectFolder}>
-        <div className="grid gap-2">
-          <label className={SESSION_SECTION_LABEL_CLASS} htmlFor={folderFieldID}>
-            {messages.sessionFolderFieldLabel}
-          </label>
-          <div className="flex flex-col gap-2 sm:flex-row">
-            <Input
-              autoFocus
-              className="flex-1"
-              disabled={isPending}
-              id={folderFieldID}
-              onChange={(event) => {
-                onChangeFolderPath(event.target.value);
-              }}
-              placeholder={messages.sessionFolderFieldPlaceholder}
-              value={folderPath}
-            />
-            <Button
-              disabled={isPending}
-              onClick={handleOpenFolderPicker}
-              size="sm"
-              tone="outline"
-              type="button"
-            >
-              {messages.browseSessionFolderButtonLabel}
-            </Button>
-          </div>
-          <input
-            {...({ directory: "", webkitdirectory: "" } as Record<string, string>)}
-            className="sr-only"
-            disabled={isPending}
-            onChange={handleSelectFolder}
-            ref={folderPickerInputRef}
-            tabIndex={-1}
-            type="file"
-          />
-        </div>
-        {dialogError ? (
-          <p className={SESSION_DIALOG_ERROR_CLASS} role="alert">
-            {dialogError.message}
-          </p>
-        ) : null}
-        <div className="flex justify-end">
-          <Button disabled={isPending} type="submit">
-            {isPending
-              ? messages.openSessionSubmitPendingLabel
-              : messages.openSessionSubmitLabel}
-          </Button>
-        </div>
-      </form>
-      {discoveredTargets.length > 0 ? (
-        <SessionTargetPicker
-          isPending={isPending}
-          messages={messages}
-          onOpenTarget={onOpenTarget}
-          targets={discoveredTargets}
-        />
-      ) : null}
-    </DialogContent>
-  );
-}
-
-function SessionTargetPicker({
-  isPending,
-  messages,
-  onOpenTarget,
-  targets,
-}: {
-  isPending: boolean;
-  messages: ReturnType<typeof getHeaderControlsMessages>;
-  onOpenTarget: (target: FactorySessionTarget) => void;
-  targets: FactorySessionTarget[];
-}) {
-  return (
-    <section aria-label={messages.targetPickerTitle} className="grid gap-3">
-      <div className="grid gap-1">
-        <p className={SESSION_SECTION_LABEL_CLASS}>{messages.targetPickerTitle}</p>
-        <p className={cn("text-sm text-af-ink/72", DASHBOARD_BODY_TEXT_CLASS)}>
-          {messages.targetPickerHint}
-        </p>
-      </div>
-      <div className={SESSION_TARGET_LIST_CLASS}>
-        {targets.map((target) => (
-          <button
-            className={SESSION_TARGET_BUTTON_CLASS}
-            disabled={isPending}
-            key={`${target.ref.kind}:${target.ref.name ?? ""}:${target.factoryDir}`}
-            onClick={() => {
-              onOpenTarget(target);
-            }}
-            type="button"
-          >
-            <span className="font-semibold text-af-ink">{target.label}</span>
-            <span className="truncate text-xs text-af-ink/58">
-              {target.project || target.factoryDir}
-            </span>
-          </button>
-        ))}
-      </div>
-    </section>
-  );
-}
-
-function sessionTabLabel(session: FactorySessionSummary): string {
-  const folderName = basename(session.folderPath) || session.project || "factory";
-  const targetName =
-    session.target.kind === "default" ? "default" : session.target.name || "named";
-  return `${folderName} / ${targetName}`;
-}
-
-function sessionCloseLabel(
-  session: FactorySessionSummary,
-  messages: ReturnType<typeof getHeaderControlsMessages>,
-): string {
-  return messages.sessionTabCloseLabelTemplate.replace(
-    "{{sessionLabel}}",
-    sessionTabLabel(session),
-  );
-}
-
-function sessionStreamStatusLabel(
-  status: DashboardStreamState["status"],
-  messages: ReturnType<typeof getHeaderControlsMessages>,
-): string {
-  if (status === "live") {
-    return messages.streamStatusLiveLabel;
-  }
-  if (status === "offline") {
-    return messages.streamStatusOfflineLabel;
-  }
-
-  return messages.streamStatusConnectingLabel;
-}
-
 function SessionTabStatusIndicator({
   status,
 }: {
@@ -613,106 +444,4 @@ function SessionTabStatusIndicator({
       ) : null}
     </span>
   );
-}
-
-function basename(path: string): string {
-  const segments = path.split(/[\\/]/).filter((segment) => segment.length > 0);
-  return segments[segments.length - 1] ?? "";
-}
-
-function sessionTabID(sessionTabsID: string, sessionID: string): string {
-  return `${sessionTabsID}-tab-${sessionDOMIDFragment(sessionID)}`;
-}
-
-function sessionPanelID(sessionTabsID: string, sessionID: string): string {
-  return `${sessionTabsID}-panel-${sessionDOMIDFragment(sessionID)}`;
-}
-
-function sessionDOMIDFragment(value: string): string {
-  return value.replace(/[^a-zA-Z0-9_-]+/g, "-");
-}
-
-function normalizeFactorySessionsError(error: unknown): FactorySessionsAPIError {
-  if (error instanceof FactorySessionsAPIError) {
-    return error;
-  }
-  return new FactorySessionsAPIError(
-    "The dashboard could not complete the factory session request.",
-    {
-      code: "INTERNAL_ERROR",
-      responseBody: error,
-    },
-  );
-}
-
-function extractSelectedFolderPath(
-  files: FileList | File[] | null,
-): string | null {
-  const firstFile = firstSelectedFile(files);
-  if (!firstFile) {
-    return null;
-  }
-
-  const absolutePath = readSelectedFilePath(firstFile);
-  if (absolutePath) {
-    return dirname(absolutePath);
-  }
-  return null;
-}
-
-async function pickDirectoryHandle(): Promise<FileSystemDirectoryHandle | null> {
-  const browserWithDirectoryPicker = globalThis as typeof globalThis & {
-    showDirectoryPicker?: () => Promise<FileSystemDirectoryHandle>;
-  };
-  if (typeof browserWithDirectoryPicker.showDirectoryPicker !== "function") {
-    return null;
-  }
-
-  try {
-    return await browserWithDirectoryPicker.showDirectoryPicker();
-  } catch (error) {
-    if (isDirectoryPickerAbortError(error)) {
-      return null;
-    }
-    throw error;
-  }
-}
-
-function firstSelectedFile(files: FileList | File[] | null): File | null {
-  if (!files) {
-    return null;
-  }
-
-  if (Array.isArray(files)) {
-    return files[0] ?? null;
-  }
-
-  return files.item(0);
-}
-
-function readSelectedFilePath(file: File): string | null {
-  const pathValue = (file as File & { path?: string }).path;
-  return typeof pathValue === "string" && isAbsolutePath(pathValue)
-    ? pathValue
-    : null;
-}
-
-function readDirectoryHandlePath(handle: FileSystemDirectoryHandle): string | null {
-  const pathValue = (handle as FileSystemDirectoryHandle & { path?: string }).path;
-  if (typeof pathValue === "string" && isAbsolutePath(pathValue)) {
-    return pathValue;
-  }
-  return null;
-}
-
-function isDirectoryPickerAbortError(error: unknown): boolean {
-  return error instanceof DOMException && error.name === "AbortError";
-}
-
-function isAbsolutePath(path: string): boolean {
-  return path.startsWith("/") || /^[a-zA-Z]:[\\/]/.test(path);
-}
-
-function dirname(path: string): string {
-  return path.replace(/[\\/][^\\/]+$/, "");
 }
